Added world helpers

+++ features/support/user_cleanup.rb
@@ -1,3 +1,5 @@
 After do
-  User.last.destroy
+  if User.present?
+    User.last.destroy
+  end
 end
